Fix unnecessary tsbuildinfo emits when getting cached dts diagnostics

+++ internal/execute/incremental/emitfileshandler.go
@@ -12,8 +12,9 @@ import (
 )
 
 type emitUpdate struct {
-	pendingKind FileEmitKind
-	result      *compiler.EmitResult
+	pendingKind        FileEmitKind
+	result             *compiler.EmitResult
+	dtsErrorsFromCache bool
 }
 
 type emitFilesHandler struct {
@@ -107,10 +108,14 @@ func (h *emitFilesHandler) emitAllAffectedFiles(options compiler.EmitOptions) *c
 				return true
 			}
 			pendingKind, _ := h.program.snapshot.affectedFilesPendingEmit.Load(path)
-			h.emitUpdates.Store(path, &emitUpdate{pendingKind: pendingKind, result: &compiler.EmitResult{
-				EmitSkipped: true,
-				Diagnostics: diagnostics.getDiagnostics(h.program.program, affectedFile),
-			}})
+			h.emitUpdates.Store(path, &emitUpdate{
+				pendingKind: pendingKind,
+				result: &compiler.EmitResult{
+					EmitSkipped: true,
+					Diagnostics: diagnostics.getDiagnostics(h.program.program, affectedFile),
+				},
+				dtsErrorsFromCache: true,
+			})
 		}
 		return true
 	})
@@ -253,10 +258,13 @@ func (h *emitFilesHandler) updateSnapshot() []*compiler.EmitResult {
 			h.program.snapshot.hasChangedDtsFile = true
 		}
 		if update, ok := h.emitUpdates.Load(file.Path()); ok {
-			if update.pendingKind == 0 {
-				h.program.snapshot.affectedFilesPendingEmit.Delete(file.Path())
-			} else {
-				h.program.snapshot.affectedFilesPendingEmit.Store(file.Path(), update.pendingKind)
+			if !update.dtsErrorsFromCache {
+				if update.pendingKind == 0 {
+					h.program.snapshot.affectedFilesPendingEmit.Delete(file.Path())
+				} else {
+					h.program.snapshot.affectedFilesPendingEmit.Store(file.Path(), update.pendingKind)
+				}
+				h.program.snapshot.buildInfoEmitPending.Store(true)
 			}
 			if update.result != nil {
 				results = append(results, update.result)
@@ -264,7 +272,6 @@ func (h *emitFilesHandler) updateSnapshot() []*compiler.EmitResult {
 					h.program.snapshot.emitDiagnosticsPerFile.Store(file.Path(), &diagnosticsOrBuildInfoDiagnosticsWithFileName{diagnostics: update.result.Diagnostics})
 				}
 			}
-			h.program.snapshot.buildInfoEmitPending.Store(true)
 		}
 	}
 	return results
